build: adjust for SVN r280013
LLVM_INCLUDE_DIRS and LLVM_LIBRARY_DIRS have been renamed to not have the S
suffix.  Remove the reference to LLVM_LIBRARY_DIRS as it was unused, update the
use of LLVM_INCLUDE_DIRS.  Swap the path in which the LLVMConfig.cmake is
searched to allow the build-tree headers to be given preference.