Fix tests for file timestamps to drop the LLVM epoch offset.
Now that Swift is not using LLVM's TimeValue (564fc6f2 and previous commit)
there is no offset from the system_clock epoch. The offset could be added
into the tests that use touch.py (so the times would not be back in 1984)
but I decided not to do that to avoid merge conflicts in the test files.