build: extract `add_swift_library` and rename to `add_swift_target_library`
Split up the `add_swift_library` by extracting the tools host handling aspect
of it.  This has a much more limited option set.  Reducing the options further
would allow us to eliminate this and replace it with a wrapper around
`add_llvm_library`.  Leave the bulk of the logic in place in
`add_swift_target_library` which can now be simplified on its own.