build: remove unnecessary flags, sink option, add import
The _Structural module is part of the standard library and has no
bearing on the compiler.  Add the option to control building the library
at the right level.

The library should not be built as the standard library as it is not
using `Builtin`.

Add a missing import to the test.